apps/shop/pyproject.toml
[project]
name = "shop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-environ>=0.11.2",
    "django-filter>=24.3",
    "django>=5.1.1",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.27.2",
    "psycopg-binary>=3.2.2",
    "psycopg>=3.2.2",
    "pytest-django>=4.9.0",
    "pytest>=8.3.3",
]

[tool.uv]
dev-dependencies = [
    "coveralls>=4.0.1",
    "factory-boy>=3.3.1",
    "faker>=30.0.0",
    "pre-commit>=3.8.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.8",
]

[tool.ruff]
line-length = 100
exclude = [
  ".env",
  "venv",
  "migrations",
]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py"]
addopts = "--last-failed --exitfirst  -v"
minversion = "7.4.4"
testpaths = ["tests"]
